Implement Control Flow Graph.
This uncovered some other issues, including the fact that the parser was
configured as static, and that the basic blocks algorithm was incorrect
(the instruction before branches, not branches themselves, are leaders).

The testing for BasicBlock should probably get a lot more attention, but
there are at least sanity-check tests for both main methods.